+++ src/eemilib/loader/pandas_loader.py
@@ -4,12 +4,13 @@ See the example TEEY in ``data/example_copper/`` for the expected file format.
 
 """
 
+import logging
 from pathlib import Path
 from typing import Any
 
 import pandas as pd
+from eemilib.loader.helper import read_header
 from eemilib.loader.loader import Loader
-from eemilib.util.constants import col_energy
 
 
 class PandasLoader(Loader):
@@ -25,45 +26,74 @@ class PandasLoader(Loader):
         sep: str = "\t",
         comment: str = "#",
     ) -> pd.DataFrame:
-        """Load and format the given emission yield files.
+        """Load and format the given emission yield file.
 
         Parameters
         ----------
-        filepath : str | pathlib.Path
+        filepath :
             Path to file holding data under study.
+        sep :
+            Column delimiter.
+        comment :
+            Comment character.
 
         Returns
         -------
-        data : pandas.DataFrame
             Structure holding the data. Has a ``Energy [eV]`` column
             holding PEs energy. And one or several columns ``theta [deg]``,
             where `theta` is the value of the incidence angle and content is
             corresponding emission yield.
 
         """
-        headers = []
-        i = 0
-        with open(filepath) as file:
-            for i, line in enumerate(file):
-                if not line.startswith(comment):
-                    headers = line.strip().split(sep)
-                    break
-        if not headers:
-            raise OSError(
-                "Error reading the given file. It seems there is no"
-                f" uncommented line? Comment character is {comment}."
-            )
-
-        headers[0] = col_energy
-        headers[1:] = [f"{float(h)} [deg]" for h in headers[1:]]
-
+        header, n_comments = read_header(filepath, sep, comment)
         df = pd.read_csv(
-            filepath, comment=comment, sep=sep, names=headers, skiprows=i + 1
+            filepath,
+            comment=comment,
+            sep=sep,
+            names=header,
+            skiprows=n_comments + 1,
         )
+        logging.info(f"Successfully loaded emission yield file(s) {filepath}")
         return df
 
     def load_emission_angle_distribution(self, *args) -> Any:
         raise NotImplementedError
 
-    def load_emission_energy_distribution(self, *args) -> Any:
-        raise NotImplementedError
+    def load_emission_energy_distribution(
+        self,
+        filepath: str | Path,
+        sep: str = "\t",
+        comment: str = "#",
+    ) -> pd.DataFrame:
+        """Load and format the given emission energy file.
+
+        Parameters
+        ----------
+        filepath :
+            Path to file holding data under study.
+        sep :
+            Column delimiter.
+        comment :
+            Comment character.
+
+        Returns
+        -------
+            Structure holding the data. Has a ``Energy [eV]`` column
+            holding emitted electrons energy. And one or several columns
+            ``theta [deg]``, where ``theta`` is the value of the incidence
+            angle and content is corresponding emission energy distribution.
+
+        """
+        header, n_comments = read_header(filepath, sep, comment)
+        df = pd.read_csv(
+            filepath,
+            comment=comment,
+            sep=sep,
+            names=header,
+            skiprows=n_comments + 1,
+        )
+        logging.info(
+            "Successfully loaded emission energy distribution file(s) "
+            f"{filepath}"
+        )
+        return df